Addressed 64 bit compiler warnings.
  * cnvts/find_service/find_service.c (mark_unknown_svc, plugin_do_run):
  Use GSIZE_TO_POINTER consistently instead casts to (void *).

  * openvassd/openvassd.c (scanner_thread): Use size_t for sizeof return
  value


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@9472 423fd1db-d629-0410-8442-d21db03e70f4

+++ openvassd/openvassd.c
@@ -484,7 +484,7 @@ scanner_thread (struct arglist *globals)
     gnutls_x509_crt_t cert;
     unsigned int cert_list_size = 0;
     const gnutls_datum_t *cert_list;
-    unsigned int x509_dname_size = sizeof (x509_dname);
+    size_t x509_dname_size = sizeof (x509_dname);
     int ret;
 
     session = ovas_get_tlssession_from_connection (soc2);
